css responsive tasarım

+++ components/edit/edit.jsx
@@ -20,28 +20,33 @@ export default function EditFeedback({ id, data, show, setShow, userId, login })
       console.log("Delete URL:", `${localUrl}/api/Opinions/${id}`);
 
       // Yorumları siliyoruz ki backend yorumlu opinionda hata vermesin
-      await deleteAllCommentsOfOpinion(id);
+     const deleteResponse = await deleteAllCommentsOfOpinion(id);
+     
       console.log("Comments deleted before deleting the opinion");
 
       // Görüşü (opinion) sil
-      const response = await fetch(`${localUrl}/api/Opinions/${id}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-          Accept: "application/json",
-        },
-        credentials: "include",
-      });
-
-      console.log(response);
-
-      if (response.ok) {
-        toast.success("Feedback deleted successfully!");
-        setShow(false);
-        window.location.href = "/";
-      } else {
-        toast.error("Failed to delete feedback.");
+      if(deleteResponse.status){
+        const response = await fetch(`${localUrl}/api/Opinions/${id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            Accept: "application/json",
+          },
+          credentials: "include",
+        });
+  
+        console.log(response);
+  
+        if (response.ok) {
+          toast.success("Feedback deleted successfully!");
+          setShow(false);
+          window.location.href = "/";
+        } else {
+          toast.error("Failed to delete feedback.");
+        }
+        
       }
+     
     } catch (error) {
       console.log(error);
       toast.error("An error occurred while deleting feedback.");
